pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mini_ros"
version = "0.1.0"
description = "Humanoid Reinforcement learning repo for World Engine"
authors = [
    { name = "World Engine Team" }
]
license = { text = "BSD-3-Clause" }
requires-python = "==3.11.*"
dependencies = [
    "pyrealsense2",
    "uvloop",
    "pyzmq>=26.3.0",
    "aiocache>=0.12.3",
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.13",
    "aiomultiprocess>=0.9.1",
    "loguru",
    "json5",
    "numpy",
    "opencv-python",
    "dynamixel_sdk",
    "feetech-servo-sdk",
    "pyyaml>=6.0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mini_ros"]

[tool.hatchling.build]
packages = ["src"]

[tool.uv.sources]
